requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
huggingface-hub
sentence-transformers
pypdf
faiss-cpu
streamlit
python-dotenv